Refactor: Move events directory into core
This commit moves the `events` directory into the `core` directory to better reflect its role as a core component of the addon. All necessary import paths have been updated to reflect this new, more logical structure.
